Move Neo event handlers to neoforge subpackage
Move ModDataFixes elsewhere, since it's platform-independent
